Benchmarks para tipos diferentes de dados

+++ ED/TP2/include/BaseDistancia.hpp
@@ -8,7 +8,7 @@
 #include <string>
 #include <ostream>
 
-typedef unsigned int distance_t;
+typedef long long distance_t;
 
 struct BaseDistancia {
   BaseDistancia() = default;
